Cleaned up file outputs in align, collapse, transcriptome, improved gtf parsing in collapse

+++ src/flair/flair_transcriptome.py
@@ -77,6 +77,10 @@ def get_args():
                              'Will be output in the final bed file but not the gtf file. '
                              'Productivity annotation is also added in the name field, '
                              'which is detailed further in the predictProductivity documentation')
+    parser.add_argument('--keep_intermediate', default=False, action='store_true',
+                        help='''specify if intermediate and temporary files are to be kept for debugging.
+                Intermediate files include: promoter-supported reads file,
+                read assignments to firstpass isoforms''')
 
     no_arguments_passed = len(sys.argv) == 1
     if no_arguments_passed:
@@ -166,7 +170,6 @@ def generateKnownSSDatabase(args, tempDir):
         if addFlag == False:
             sys.stderr.write('\nERROR Added no extra junctions from {}\n\n'.format(args.shortread))
             sys.exit(1)
-    knownSS = dict()
 
     sys.stderr = sys.__stderr__
 
@@ -176,18 +179,17 @@ def generateKnownSSDatabase(args, tempDir):
         sys.exit(1)
 
     annotationFiles = dict()
-    chrom2 = set()
-    for chrom, data in juncs.items():
-        if len(data) > 0:
-            chrom2.add(chrom)
-            annotationFiles[chrom] = os.path.join(tempDir, "%s_known_juncs.bed" % chrom)
-            with open(os.path.join(tempDir, "%s_known_juncs.bed" % chrom), "w") as out:
+    for chrom in chromosomes:
+        annotationFiles[chrom] = os.path.join(tempDir, "%s_known_juncs.bed" % chrom)
+        with open(os.path.join(tempDir, "%s_known_juncs.bed" % chrom), "w") as out:
+            if chrom in juncs:
+                data = juncs[chrom]
                 sortedData = sorted(list(data.keys()), key=lambda item: item[0])
                 for k in sortedData:
                     annotation = data[k]
                     c1, c2, strand = k
                     print(chrom, c1, c2, annotation, ".", strand, sep="\t", file=out)
-    return chrom2, annotationFiles
+    return chromosomes, annotationFiles
 
 
 def correctsingleread(bedread, intervalTree, junctionBoundaryDict):
@@ -967,8 +969,8 @@ def getbedgtfoutfrominfo(endinfo, chrom, strand, juncs, gene, genome):
 
 
 def combineannotnovelwriteout(args, genetojuncstoends, genome):
-    with open(args.output + '.isoforms.bed', 'w') as isoout, open(args.output + '.read.map.txt', 'w') as mapout, open(
-            args.output + '.isoforms.gtf', 'w') as gtfout, open(args.output + '.isoforms.fa', 'w') as seqout:
+    with open(args.output + '.isoforms.bed', 'w') as isoout, open(args.output + '.isoform.read.map.txt', 'w') as mapout, open(
+            args.output + '.isoforms.gtf', 'w') as gtfout, open(args.output + '.isoforms.fa', 'w') as seqout, open(args.output + '.isoform.counts.txt', 'w') as countsout:
         for gene in genetojuncstoends:
             gtflines, tstarts, tends = [], [], []
             for chrom, strand, juncs in genetojuncstoends[gene]:
@@ -1002,6 +1004,7 @@ def combineannotnovelwriteout(args, genetojuncstoends, genome):
                     tends.append(isoinfo[1])
                     gtflines.extend(gtffortranscript)
                     mapout.write(iso + '_' + gene + '\t' + ','.join(isoinfo[3]) + '\n')
+                    countsout.write(iso + '_' + gene + '\t' + str(len(isoinfo[3])) + '\n')
                     seqout.write('>' + iso + '_' + gene + '\n')
                     seqout.write(tseq + '\n')
             gtflines.insert(0, [chrom, 'FLAIR', 'gene', min(tstarts) + 1, max(tends), '.', gtflines[0][6], '.',
@@ -1063,8 +1066,9 @@ def runcollapsebychrom(listofargs):
             tempprefix + '.novelisos.read.map.txt', 'w') as mapout:
             pass
     genome.close()
-    for f in temptoremove:
-        os.remove(f)
+    if not args.keep_intermediate:
+        for f in temptoremove:
+            os.remove(f)
 
 
 def collapsefrombam():
@@ -1137,7 +1141,8 @@ def collapsefrombam():
                              ['.firstpass.reallyunfiltered.bed', '.firstpass.unfiltered.bed', '.firstpass.bed', '.novelisos.counts.tsv',
                               '.novelisos.read.map.txt'])
 
-    shutil.rmtree(tempDir)
+    if not args.keep_intermediate:
+        shutil.rmtree(tempDir)
 
     if not args.noaligntoannot:
         genetojuncstoends = processdetectedisos(args, args.output + '.matchannot.read.map.txt',
@@ -1148,6 +1153,19 @@ def collapsefrombam():
                                             genetojuncstoends)
 
     combineannotnovelwriteout(args, genetojuncstoends, genome)
+    if not args.keep_intermediate:
+        files_to_remove = ['.firstpass.reallyunfiltered.bed',
+                           '.firstpass.unfiltered.bed',
+                           '.firstpass.bed',
+                           '.novelisos.counts.tsv',
+                           '.novelisos.read.map.txt']
+        if not args.noaligntoannot:
+            files_to_remove += ['.matchannot.bed',
+                           '.matchannot.counts.tsv',
+                           '.matchannot.read.map.txt']
+        for f in files_to_remove:
+            os.remove(args.output + f)
+
     if args.predictCDS:
         prodcmd = ('predictProductivity',
                    '-i', args.output+'.isoforms.bed',
